chatbot_ventas/requirements.txt
google-generativeai
python-dotenv
pandas
